requirements.txt
PyQt6==6.7.0
requests==2.31.0
python-dotenv==1.0.1
openai==1.3.7
python-telegram-bot==20.6
telethon==1.34.0
psutil==5.9.5
pyautogui==0.9.54
speechrecognition==3.10.0
pillow==10.0.1
darkdetect==0.8.0
speedtest-cli==2.1.3
replicate
gradio_client

# Сборка exe
pyinstaller>=6.0.0

# Voice (optional)
vosk==0.3.45

# NeuTTS Air (optional)
# NeuTTS isn't published on PyPI. Install from GitHub when needed:
#   pip install "git+https://github.com/neuphonic/neutts.git"
#   pip install soundfile
